feat(visits): add visit count endpoints for admin; remove unused counts from client routes
