+++ src/controller/visits.admin.controller.js
@@ -1,0 +1,36 @@
+import { Visit } from "../model/visit.model.js";
+
+
+export const getAllVisitsCount = async (_, res, next) => {
+
+    try {
+        const totalVisits = await Visit.countDocuments();
+
+        return res.status(200).json({
+            status: true,
+            message: "Total visits count fetched successfully",
+            total: totalVisits
+        });
+    }
+
+    catch (error) {
+        next(error);
+    }
+}
+
+export const getPendingVisitsCount = async (_, res, next) => {
+
+    try {
+        const totalVisits = await Visit.countDocuments({ status: "pending" });
+
+        return res.status(200).json({
+            status: true,
+            message: "Total pending visits count fetched successfully",
+            total: totalVisits
+        });
+    }
+
+    catch (error) {
+        next(error)
+    }
+}
